requirements.txt
django == 3.1.2
pillow
django-crispy-forms== 1.13.0
django-rest-framework
social-auth-app-django
django-pwa
django-cleanup
python-jose >= 3.0.0
django-environ==0.8.1
django-choices==1.7.1